Fix commander broadcast for commanders who left the channel
Since temporary groups remain even after the user has left the
channel (eg: to the AFK channel), that user would still be
broadcasted as an active commander. With this commit, only
users who have the active commander role AND are in the channel
they have that role in will be broadcasted.