Rewrite control flow
The inner TS3 connection uses a lot of
exception handling to do control flow.
For example, looking for a user that is not present
causes an exception with specific code to signal their absence.
This commits wraps the exception handling in an own callback
when calling the ts3exec (formerly tsc) method of the
ThreadsafeTS3Connection.
The method now returns a pair of values: the first one being
the result of the call, the second one being the result of the
exception handler. Returning None from that handler signals that
everything went smoothly, while having anything but None signals
the occurrence of an exception (may have been handled already, but
should still be signaled in most cases to do proper control flow).

This changes the control flow from unstructured exception catching
to functional Maybe-style error handling.